- Corrected char-server parameters so that the subnet config file is the third parameter, not the second (which was conflicting with inter_athena.conf's argument), thanks to foobar. - Fixed TK and Soul Linker's position of their job entry in exp2.txt, thanks to Coltaro. - Fixed NPC_GRANDDARKNESS targetting the enemy instead of yourself in mob_skill_db. - Fixed Grandcross's hit and range entries in the skill_db (it had hit 5? that wasn't even a valid value from the description) - PR_BENEDICTIO now ignores mdef - Removed the status_isdead checks from status_check_skilluse, they are now done only on the skill_castend_* functions. - Fixed NPC_SELFDESTRUCTION not doing the correct damage to oneself. - Fixed chatlog's x/y column capacity.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5584 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ sql-files/logs.sql
@@ -214,8 +214,8 @@ CREATE TABLE `chatlog` (
   `src_charid` int(11) NOT NULL default '0',
   `src_accountid` int(11) NOT NULL default '0',
   `src_map` varchar(17) NOT NULL default 'prontera.gat',
-  `src_map_x` tinyint(4) NOT NULL default '0',
-  `src_map_y` tinyint(4) NOT NULL default '0',
+  `src_map_x` smallint(4) NOT NULL default '0',
+  `src_map_y` smallint(4) NOT NULL default '0',
   `dst_charname` varchar(25) NOT NULL default '',
   `message` varchar(150) NOT NULL default '',
   PRIMARY KEY  (`id`)
